backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
aiofiles
Pillow
python-multipart
google-cloud-vision